Cleaning up file handling

+++ AccretionTools.py
@@ -1355,7 +1355,7 @@ def add_gas_particle_data(base_halo_data,accdata_path,datasets=None):
         print(type(gas_IDs_in_snap1))
         print(gas_IDs_in_snap1)
         
-        if np.size(gas_IDs_in_snap1)==1 and type(gas_IDs_in_snap1)==float:
+        if np.size(gas_IDs_in_snap1)==1 and type(gas_IDs_in_snap1)==np.float16:
             for dataset in datasets:
                 ihalo_datasets_inflow[f'snap1_{dataset}']=np.nan
                 ihalo_datasets_outflow[f'snap1_{dataset}']=np.nan
